Fix big buffers (4GB+) overflow in ttnn codebase and metal slow dispatch (#22731)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-metal/issues/22336)

### Problem description
ttnn.from_torch / to_torch doesn't really work well for huge tensors
with sizes not fitting in uint32_t. This PR is first step to fix that

### What's changed
Fix integer overflow in TTNN code base:
- tilizing
- padding

Fix integer overflow in metal code base for slow dispatch
- Most changed in `tt_metal/api/tt-metalium/buffer.hpp`

Added c++ test for tilizing of 4Gb+ buffer:
- `tests/tt_metal/tt_metal/api/test_tilize_untilize.cpp`
- It's turned off by default, since it takes a few minutes to run. If
you need to run, pass `--gtest_also_run_disabled_tests` to gtest binary

Added python test for moving to device and then back > 4Gb of data:
- `tests/ttnn/stress_tests/test_create_buffer_bigger_than_4gb.py`
- Add `slow` mark for ttnn pytests to mark slow tests, that won't run by
default, unless `--runslow` is passed (see `tests/ttnn/conftest.py`)
- This tests takes a ton of time, since slow dispatch readback is
extremely slow

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes --
https://github.com/tenstorrent/tt-metal/actions/runs/15819570114/job/44589848819
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI with demo tests passes (if applicable) --
https://github.com/tenstorrent/tt-metal/actions/runs/15821968999
- [ ] [Model
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml)
CI passes (if applicable)
- [ ] [Device performance
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml)
CI passes (if applicable)
- [ ] (For models and ops writers) [Single-card demo
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml)
CI passes (if applicable) See [recommended dev
flow](https://github.com/tenstorrent/tt-metal/blob/main/models/MODEL_ADD.md#a-recommended-dev-flow-on-github-for-adding-new-models).
- [x] New/Existing tests provide coverage for changes

+++ ttnn/core/tensor/tensor_impl.cpp
@@ -1336,7 +1336,7 @@ Tensor pad(
     const auto input_strides = tensor.strides();
 
     auto pad = [&input_padded_shape, &output_padded_shape, &input_tensor_start, &pad_value_](const auto& input_buffer) {
-        const int rank = input_padded_shape.rank();
+        const size_t rank = input_padded_shape.rank();
 
         auto output_buffer = std::vector<T>(output_padded_shape.volume());
         std::fill(output_buffer.begin(), output_buffer.end(), pad_value_);
@@ -1347,7 +1347,9 @@ Tensor pad(
 
         if (rank == 1) {
             std::memcpy(
-                output_buffer.data() + input_tensor_start[0], input_buffer.begin(), input_padded_shape[0] * sizeof(T));
+                output_buffer.data() + input_tensor_start[0],
+                input_buffer.begin(),
+                static_cast<size_t>(input_padded_shape[0]) * sizeof(T));
             return output_buffer;
         }
 
@@ -1366,17 +1368,17 @@ Tensor pad(
 
             for (int i = 0; i < rank - 1; ++i) {
                 input_idx += coords[i] * input_strides[i];
-                output_idx += (coords[i] + input_tensor_start[i]) * output_strides[i];
+                output_idx += (coords[i] + static_cast<size_t>(input_tensor_start[i])) * output_strides[i];
             }
 
             // Add offset (left padding) for the innermost dimension
-            output_idx += input_tensor_start[rank - 1] * output_strides[rank - 1];
+            output_idx += static_cast<size_t>(input_tensor_start[rank - 1]) * output_strides[rank - 1];
 
             // Copy entire input row with memcpy
             std::memcpy(
                 output_buffer.data() + output_idx,
                 input_buffer.begin() + input_idx,
-                input_padded_shape[rank - 1] * sizeof(T));
+                static_cast<size_t>(input_padded_shape[rank - 1]) * sizeof(T));
 
             // Increment coordinates (from right to left), ignore last dimension
             processed_all_coords = true;
